Stop crashes when connecting to autotracking
When connecting to auto tracking if the device name is longer than 20 characters e.g. "emunwa://" or "luabridge://" the program will crash with "Invalid call. Nonexistent function 'substring' in base 'String'.". Changing this to the correct .substr() allows for autotracking to be used. (im assuming you tested in retro arch which is "ra://")